+++ src/pytest_mysql/__init__.py
@@ -20,7 +20,7 @@
 
 import logging
 
-__version__ = '1.1.1'
+__version__ = '2.0.0'
 
 
 logger = logging.getLogger(__name__)